Move CanvasScalerSetting class into CanvasScalerReplacer.cs
