+++ Assets/Project/Scripts/Editor/CanvasScalerReplacer.cs
@@ -2,10 +2,20 @@
 using UnityEditor;
 using UnityEngine.UI;
 using UnityEditor.SceneManagement;
-using Project.Scripts.Settings;
 
 namespace Project.Scripts.Editor
 {
+    public class CanvasScalerSetting : ScriptableObject
+    {
+        [SerializeField] private Vector2 referenceResolution;
+
+        [SerializeField] [Range(0, 1)] private float matchWidthOrHeight;
+
+        public Vector2 ReferenceResolution => referenceResolution;
+
+        public float MatchWidthOrHeight => matchWidthOrHeight;
+    }
+
     public class CanvasScalerReplacer : UnityEditor.EditorWindow
     {
         /// <summary>
