--- Assets/Project/Scripts/Settings.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 63be1c3f1dce4420a88dd2b369ab7c3b
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
