--- Assets/Project/Scripts/Settings/CanvasScalerSetting.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c7aea47a0671b467192f95c3f82b0fdc
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
